test build 13.0

+++ utils/github/install-macos.sh
@@ -23,17 +23,17 @@ rm -rf \
 
 brew tap copyq/kde utils/github/homebrew/
 
-qt_minimum_target=12.0
-curl -O https://raw.githubusercontent.com/Homebrew/homebrew-core/refs/heads/master/Formula/q/qt.rb
-sed -i.bak \
-"s|-DCMAKE_OSX_DEPLOYMENT_TARGET=#{MacOS.version}\.0|\
--DCMAKE_OSX_DEPLOYMENT_TARGET=${qt_minimum_target}|g" \
-qt.rb
-mv qt.rb.bak qt.rb
-brew install --build-from-source ./qt.rb
-brew uninstall vulkan-headers vulkan-loader molten-vk node pkgconf
+if [[ $(uname -m) == 'x86_64' ]]; then
+    qt_minimum_target=13.0
+    curl -O https://raw.githubusercontent.com/Homebrew/homebrew-core/refs/heads/master/Formula/q/qt.rb
+    sed -i.bak "s|-DCMAKE_OSX_DEPLOYMENT_TARGET=#{MacOS.version}\.0|-DCMAKE_OSX_DEPLOYMENT_TARGET=${qt_minimum_target}|g" qt.rb
+    mv qt.rb.bak qt.rb
+    brew install --build-from-source --formula ./qt.rb
+    brew uninstall vulkan-headers vulkan-loader molten-vk node pkgconf
+else
+    brew install qt@6
+fi
 
-# brew install qt@6
 brew install --verbose \
     copyq/kde/kf6-knotifications \
     copyq/kde/kf6-kstatusnotifieritem
